backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-Bcrypt==1.0.1
plaid_python==9.1.0
python-dotenv==0.15.0
marshmallow==3.19.0
